requirements.txt
# ── Web Framework ─────────────────────────────────────────────
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# ── ML ────────────────────────────────────────────────────────
torch==2.4.1
transformers==4.46.1
numpy==1.26.4
scikit-learn==1.5.2

# ── Database ──────────────────────────────────────────────────
sqlalchemy==2.0.36

# ── Scraping ──────────────────────────────────────────────────
trafilatura==1.12.2
httpx==0.27.2
beautifulsoup4==4.12.3

# ── Utilities ─────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0

# ── Rate Limiting ─────────────────────────────────────────────
slowapi==0.1.9

# ── Logging ───────────────────────────────────────────────────
python-json-logger==2.0.7

# ── Production Server ─────────────────────────────────────────
gunicorn==23.0.0
